services/bq-ingest/requirements.txt
Flask==3.0.3
requests==2.32.3
google-cloud-bigquery==3.41.0
gunicorn==22.0.0
google-cloud-secret-manager==2.17.0
google-genai==1.47.0
tenacity==8.5.0
